# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/stores/{id}": {
            "get": {
                "tags": [
                    "Store"
                ],
                "summary": "Display the specified Store",
                "description": "Get Store by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Store",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Store"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Store"
                ],
                "summary": "Update the specified Store in storage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Store",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Store"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Store"
                ],
                "summary": "Remove the specified Store from storage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Store",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Store": {
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "latitud": {
                    "type": "number",
                    "format": "float"
                },
                "longitud": {
                    "type": "number",
                    "format": "float"
                },
                "city_id": {
                    "type": "integer"
                },
                "owner_id": {
                    "type": "integer"
                },
                "facebook": {
                    "type": "string"
                },
                "instagram": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
    }
}